requirements.txt
csv==1.0
imageio==2.9.0
imagesize==1.2.0
imbalanced-learn==0.7.0
matplotlib==3.3.3
mlxtend==0.18.0
numpy==1.19.4
opencv-python==4.5.1.48
openpyxl==3.0.6
pandas==1.1.4
Pillow==8.0.1
pyradiomics==3.0.1
PyWavelets==1.1.1
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.1
SimpleITK==2.0.1
six==1.15.0
xlrd==1.2.0
XlsxWriter==1.3.7
ReliefF==0.1.2
#jsons==2.0.9
